requirements.txt
line-bot-sdk
flask
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
requests==2.18.1
six==1.10.0
urllib3==1.21.1
Werkzeug==0.12.2
mysql-connector-python==8.0.11